+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Features
 
 * Add `Castor\Event\ContextCreatedEvent` to allow updating the context after it is created
+* Add `run_phar()` function to run a phar file in all contexts
 
 ## 0.17.1 (2024-05-31)